+++ lib/deskbot/types.rb
@@ -26,13 +26,5 @@ module Deskbot
       "up",
       "down"
     )
-
-    EventType = Types::Coercible::Symbol.enum(
-      :key_down,
-      :key_up,
-      :mouse_down,
-      :mouse_up,
-      :mouse_move
-    )
   end
 end